+++ core/src/main/python/discover.py
@@ -50,6 +50,7 @@ from wlsdeploy.util import tool_exit
 from wlsdeploy.util.cla_utils import CommandLineArgUtil
 from wlsdeploy.util.model import Model
 from wlsdeploy.util.weblogic_helper import WebLogicHelper
+from wlsdeploy.util import target_configuration_helper
 
 wlst_helper.wlst_functions = globals()
 
@@ -102,7 +103,7 @@ def __process_args(args):
     combined_arg_map.update(required_arg_map)
     return model_context_helper.create_context(_program_name, combined_arg_map)
 
-def __process_target_arg(optional_arg_map, required_arg_map):
+def __process_target_arg(optional_arg_map):
 
     _method_name = '__process_target_arg'
 
@@ -431,7 +432,7 @@ def __check_and_customize_model(model, model_context, aliases, injector):
         if model_context.is_target_k8s():
             validation_method = model_context.get_target_configuration()['validation_method']
             model_context.set_validation_method(validation_method)
-            generate_k8s_script(model_context.get_kubernetes_variable_file(), cache)
+            target_configuration_helper.generate_k8s_script(model_context.get_kubernetes_variable_file(), cache)
             cache.clear()
 
     variable_injector = VariableInjector(_program_name, model.get_model(), model_context,
@@ -451,59 +452,6 @@ def __check_and_customize_model(model, model_context, aliases, injector):
     except ValidateException, ex:
         __logger.warning('WLSDPLY-06015', ex.getLocalizedMessage(), class_name=_class_name, method_name=_method_name)
     return model
-
-def generate_k8s_script(file_location, token_dictionary):
-    if file_location:
-        NL = '\n'
-        par_dir = os.path.abspath(os.path.join(file_location,os.pardir))
-        k8s_file = os.path.join(par_dir, "create_k8s_secrets.sh")
-        k8s_create_script_handle = open(k8s_file, 'w')
-        k8s_create_script_handle.write('#!/bin/bash')
-        k8s_create_script_handle.write(NL)
-        k8s_create_script_handle.write('set -eu')
-        k8s_create_script_handle.write(NL)
-        k8s_create_script_handle.write('NAMESPACE=default')
-        k8s_create_script_handle.write(NL)
-        k8s_create_script_handle.write('DOMAIN_UID=domain1')
-        k8s_create_script_handle.write(NL)
-        k8s_create_script_handle.write('ADMIN_USER=wlsAdminUser')
-        k8s_create_script_handle.write(NL)
-        k8s_create_script_handle.write('ADMIN_PWD=wlsAdminPwd')
-        k8s_create_script_handle.write(NL)
-        k8s_create_script_handle.write('function create_k8s_secret {')
-        k8s_create_script_handle.write(NL)
-        k8s_create_script_handle.write('kubectl -n $NAMESPACE delete secret ${DOMAIN_UID}-$1 --ignore-not-found')
-        k8s_create_script_handle.write(NL)
-        k8s_create_script_handle.write('kubectl -n $NAMESPACE create secret generic ${DOMAIN_UID}-$1 ' +
-                                       '--from-literal=$2=$3')
-        k8s_create_script_handle.write(NL)
-        k8s_create_script_handle.write('kubectl -n $NAMESPACE label secret ${DOMAIN_UID}-$1 ' +
-                                       'weblogic.domainUID=${DOMAIN_UID}')
-        k8s_create_script_handle.write(NL)
-        k8s_create_script_handle.write('}')
-        k8s_create_script_handle.write(NL)
-        k8s_create_script_handle.write(NL)
-        for property_name in token_dictionary:
-            if property_name in [ 'AdminUserName', 'AdminPassword']:
-                continue
-            secret_names = property_name.lower().replace('.', '-').split('-')
-            command_string = "create_k8s_secret %s %s %s " %( '-'.join(secret_names[:-1]), secret_names[-1],
-                                                              "<changeme>")
-            k8s_create_script_handle.write(command_string)
-            k8s_create_script_handle.write(NL)
-
-        k8s_create_script_handle.write(NL)
-        k8s_create_script_handle.write("kubectl -n $NAMESPACE delete secret ${DOMAIN_UID}-weblogic-credential "
-                                       + "--ignore-not-found")
-        k8s_create_script_handle.write(NL)
-        k8s_create_script_handle.write("kubectl -n $NAMESPACE create secret generic "
-                                       +  "${DOMAIN_UID}-weblogic-credential "
-                                       +   "--from-literal=username=${ADMIN_USER} --from-literal=password=${ADMIN_PWD}")
-        k8s_create_script_handle.write(NL)
-        k8s_create_script_handle.write('kubectl -n $NAMESPACE label secret ${DOMAIN_UID}-weblogic-credential ' +
-                                       'weblogic.domainUID=${DOMAIN_UID}')
-        k8s_create_script_handle.write(NL)
-        k8s_create_script_handle.close()
 
 
 def __log_and_exit(model_context, exit_code, class_name, method_name):
